+++ maiko-macros/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.3 Unreleased
+
+### Added
+
+- `Label` derive macro
+
+---
 # [0.2.2](https://github.com/ddrcode/maiko/compare/v0.2.1...v0.2.2) (February 4th, 2026)
 
 No changes. Version bumped to be aligned with the main library.